Move migrate cmd
